requirements.txt
docker-compose>=1.7
ansi2html>=1.1
